+++ pljit/Milestone5/ASTEvaluator.cpp
@@ -1,7 +1,7 @@
 #include "ASTEvaluator.h"
 #include <utility>
 namespace semantic{
-std::optional<double> ASTEvaluator::evaluateFunction(std::vector<long> arg,semantic::ASTNode& node){
+std::optional<double> ASTEvaluator::evaluateFunction(std::vector<double> arg,semantic::ASTNode& node){
     arguments = std::move(arg);
     return node.acceptEvaluation(*this);
 }
