+++ test/TestEvaluation.cpp
@@ -15,7 +15,7 @@ using SemanticAnalyzer = semantic::SemanticAnalyzer;
 using DeadCodeEliminationPass = semantic::DeadCodeEliminationPass;
 using ConstantPropagationPass = semantic::ConstantPropagationPass;
 
-static auto setupWithOptimizationShort(const std::vector<long>& args, const std::string_view& input){
+static auto setupWithOptimizationShort(const std::vector<double>& args, const std::string_view& input){
     SourceCodeManager manager(input);
     Tokenizer tokenizer = Tokenizer(manager);
 
@@ -31,7 +31,7 @@ static auto setupWithOptimizationShort(const std::vector<long>& args, const std:
     constantPropagationPass.optimize(*astNode);
     return evaluator.evaluateFunction(args,*astNode);
 }
-static auto setupWithOptimization(const std::vector<long>& args, const std::string_view& input){
+static auto setupWithOptimization(const std::vector<double>& args, const std::string_view& input){
     SourceCodeManager manager(input);
     Tokenizer tokenizer = Tokenizer(manager);
 
@@ -48,7 +48,7 @@ static auto setupWithOptimization(const std::vector<long>& args, const std::stri
 }
 
 TEST(TestEvaluation, ConstantPropagation){
-    auto args = std::vector<long>();
+    auto args = std::vector<double>();
     auto astNode = setupWithOptimization(args,"CONST a = 1, b = 2; BEGIN RETURN a+b END.");
     //test statement child nodes
     auto astChild = astNode->getChildren();
@@ -84,7 +84,7 @@ TEST(TestEvaluation, ConstantPropagation){
     EXPECT_EQ(result.value(), 108);
 }
 TEST(TestEvaluation, DeadCodeElimination){
-    auto args = std::vector<long>();
+    auto args = std::vector<double>();
     auto astNode = setupWithOptimization(args,"VAR a,b,c,d,e,f,z,g,i,j,ab,aaa ; BEGIN ab := 1 ;RETURN ab;ab := 1 ;ab := 1 ;ab := 1  END.");
     ASSERT_EQ(astNode->getType(), semantic::ASTNode::FunctionDefinition);
     //check we only have 2 statements in the optimized tree
@@ -96,31 +96,31 @@ TEST(TestEvaluation, DeadCodeElimination){
     EXPECT_EQ(compoundStatementChild.size() , 2);
 }
 TEST(TestEvaluation,MixedOptimization){
-    auto args = std::vector<long>();
+    auto args = std::vector<double>();
     auto astNode = setupWithOptimization(args,"VAR a, b; BEGIN a := 1; b := 1; RETURN a *(-b + 55 * (1-(-1))) ;a := 1; b := 1;a := 1; b := 1 END.");
     semantic::ASTEvaluator evaluator = semantic::ASTEvaluator();
     auto result = evaluator.evaluateFunction(args,*astNode);
     EXPECT_EQ(result.value(), 109);
 }
 TEST(TestEvaluation, ValidEvaluation){
-    auto result = setupWithOptimizationShort(std::vector<long>(),"VAR a,b,c,d,e,f,z,g,i,j,ab,aaa ; BEGIN ab := 1 ;RETURN ab END.");
+    auto result = setupWithOptimizationShort(std::vector<double>(),"VAR a,b,c,d,e,f,z,g,i,j,ab,aaa ; BEGIN ab := 1 ;RETURN ab END.");
     EXPECT_EQ(result.value(), 1);
-    result = setupWithOptimizationShort(std::vector<long>(),"VAR a,b,c ; BEGIN a := 1; b:=6; c:=2 ;RETURN a+b*(c+2) END.");
+    result = setupWithOptimizationShort(std::vector<double>(),"VAR a,b,c ; BEGIN a := 1; b:=6; c:=2 ;RETURN a+b*(c+2) END.");
     EXPECT_EQ(result.value(), 25);
-    result = setupWithOptimizationShort(std::vector<long>{50,2},"PARAM a,b; VAR c; BEGIN c:=2+b ;RETURN a+b*c END.");
+    result = setupWithOptimizationShort(std::vector<double>{50,2},"PARAM a,b; VAR c; BEGIN c:=2+b ;RETURN a+b*c END.");
     EXPECT_EQ(result.value(), 58);
-    result = setupWithOptimizationShort(std::vector<long>{50,2},"PARAM a,b; CONST c=1; BEGIN a:=a+c ;RETURN a+b*c END.");
+    result = setupWithOptimizationShort(std::vector<double>{50,2},"PARAM a,b; CONST c=1; BEGIN a:=a+c ;RETURN a+b*c END.");
     EXPECT_EQ(result.value(), 53);
 }
 TEST(TestEvaluation, InvalidEvaluation){
     std::cout << "Testing divide by zero statement:" << std::endl;
-    auto result = setupWithOptimizationShort(std::vector<long>(),"VAR a,b,c,d,e,f,z,g,i,j,ab,aaa ; BEGIN ab := 1/0 ;RETURN ab END.");
+    auto result = setupWithOptimizationShort(std::vector<double>(),"VAR a,b,c,d,e,f,z,g,i,j,ab,aaa ; BEGIN ab := 1/0 ;RETURN ab END.");
     EXPECT_EQ(result.has_value(), false);
-    result = setupWithOptimizationShort(std::vector<long>(),"VAR a, b; BEGIN a := 1; b := 1; RETURN (a+b * 2/(4-4))END.");
+    result = setupWithOptimizationShort(std::vector<double>(),"VAR a, b; BEGIN a := 1; b := 1; RETURN (a+b * 2/(4-4))END.");
     EXPECT_EQ(result.has_value(), false);
-    result = setupWithOptimizationShort(std::vector<long>{0,2},"PARAM a,b; VAR c; BEGIN c:=2+b ;RETURN b*c/a END.");
+    result = setupWithOptimizationShort(std::vector<double>{0,2},"PARAM a,b; VAR c; BEGIN c:=2+b ;RETURN b*c/a END.");
     EXPECT_EQ(result.has_value(), false);
-    result = setupWithOptimizationShort(std::vector<long>{50,2},"PARAM a,b; CONST c=2; BEGIN a:=a+c ;RETURN a/(b-c) END.");
+    result = setupWithOptimizationShort(std::vector<double>{50,2},"PARAM a,b; CONST c=2; BEGIN a:=a+c ;RETURN a/(b-c) END.");
     EXPECT_EQ(result.has_value(), false);
     std::cout << "=========================================================" << std::endl;
 }
